+++ android/app/src/main/kotlin/deckers/thibault/aves/metadata/MultiPage.kt
@@ -11,6 +11,7 @@ import android.os.ParcelFileDescriptor
 import android.util.Log
 import com.adobe.internal.xmp.XMPMeta
 import com.drew.imaging.jpeg.JpegSegmentType
+import com.drew.lang.SequentialByteArrayReader
 import com.drew.metadata.exif.ExifDirectoryBase
 import com.drew.metadata.exif.ExifIFD0Directory
 import com.drew.metadata.xmp.XmpDirectory
@@ -37,6 +38,8 @@ import androidx.exifinterface.media.ExifInterfaceFork as ExifInterface
 object MultiPage {
     private val LOG_TAG = LogUtils.createTag<MultiPage>()
 
+    // TODO TLAD more generic support, (e.g. 0x00000014 + `ftyp` + `qt  `)
+    // atom length (variable, e.g. `0x00000018`) + atom type (`ftyp`) + type (variable, e.g. `mp42`, `qt`)
     private val heicMotionPhotoVideoStartIndicator = byteArrayOf(0x00, 0x00, 0x00, 0x18) + "ftypmp42".toByteArray()
 
     // page info
@@ -82,6 +85,22 @@ object MultiPage {
         }
         extractor.release()
         return tracks
+    }
+
+    fun isHeicSefdMotionPhoto(context: Context, uri: Uri): Boolean {
+        Mp4ParserHelper.getSamsungSefd(context, uri)?.let { (_, bytes) ->
+            val reader = SequentialByteArrayReader(bytes).apply {
+                isMotorolaByteOrder = false
+            }
+            val start = reader.uInt16
+            val tag = reader.uInt16
+            if (start == 0 && tag == Mp4ParserHelper.SEFD_EMBEDDED_VIDEO_TAG) {
+                val nameSize = reader.uInt32
+                val name = reader.getString(nameSize.toInt())
+                return name == Mp4ParserHelper.SEFD_MOTION_PHOTO_NAME
+            }
+        }
+        return false
     }
 
     private fun getJpegMpfPrimaryRotation(context: Context, uri: Uri, sizeBytes: Long): Int {
@@ -245,40 +264,38 @@ object MultiPage {
 
     fun getMotionPhotoPages(context: Context, uri: Uri, mimeType: String, sizeBytes: Long): ArrayList<FieldMap> {
         val pages = ArrayList<FieldMap>()
-        getMotionPhotoVideoSize(context, uri, mimeType, sizeBytes)?.let { videoSizeBytes ->
-            getTrailerVideoInfo(context, uri, fileSizeBytes = sizeBytes, videoSizeBytes = videoSizeBytes)?.let { videoInfo ->
-                // set the original image as the first and default track
-                var pageIndex = 0
-                pages.add(
-                    hashMapOf(
-                        KEY_PAGE to pageIndex++,
-                        KEY_MIME_TYPE to mimeType,
-                        KEY_IS_DEFAULT to true,
-                    )
+        getMotionPhotoVideoInfo(context, uri, mimeType, sizeBytes)?.let { videoInfo ->
+            // set the original image as the first and default track
+            var pageIndex = 0
+            pages.add(
+                hashMapOf(
+                    KEY_PAGE to pageIndex++,
+                    KEY_MIME_TYPE to mimeType,
+                    KEY_IS_DEFAULT to true,
                 )
-                // add video tracks from the appended video
-                videoInfo.getString(MediaFormat.KEY_MIME)?.let { mime ->
-                    if (MimeTypes.isVideo(mime)) {
-                        val page: FieldMap = hashMapOf(
-                            KEY_PAGE to pageIndex++,
-                            KEY_MIME_TYPE to MimeTypes.MP4,
-                            KEY_IS_DEFAULT to false,
-                        )
-                        videoInfo.getSafeInt(MediaFormat.KEY_WIDTH) { page[KEY_WIDTH] = it }
-                        videoInfo.getSafeInt(MediaFormat.KEY_HEIGHT) { page[KEY_HEIGHT] = it }
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                            videoInfo.getSafeInt(MediaFormat.KEY_ROTATION) { page[KEY_ROTATION_DEGREES] = it }
-                        }
-                        videoInfo.getSafeLong(MediaFormat.KEY_DURATION) { page[KEY_DURATION] = it / 1000 }
-                        pages.add(page)
+            )
+            // add video tracks from the appended video
+            videoInfo.getString(MediaFormat.KEY_MIME)?.let { mime ->
+                if (MimeTypes.isVideo(mime)) {
+                    val page: FieldMap = hashMapOf(
+                        KEY_PAGE to pageIndex++,
+                        KEY_MIME_TYPE to MimeTypes.MP4,
+                        KEY_IS_DEFAULT to false,
+                    )
+                    videoInfo.getSafeInt(MediaFormat.KEY_WIDTH) { page[KEY_WIDTH] = it }
+                    videoInfo.getSafeInt(MediaFormat.KEY_HEIGHT) { page[KEY_HEIGHT] = it }
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                        videoInfo.getSafeInt(MediaFormat.KEY_ROTATION) { page[KEY_ROTATION_DEGREES] = it }
                     }
+                    videoInfo.getSafeLong(MediaFormat.KEY_DURATION) { page[KEY_DURATION] = it / 1000 }
+                    pages.add(page)
                 }
             }
         }
         return pages
     }
 
-    fun getMotionPhotoVideoSize(context: Context, uri: Uri, mimeType: String, sizeBytes: Long): Long? {
+    fun getTrailerVideoSize(context: Context, uri: Uri, mimeType: String, sizeBytes: Long): Long? {
         if (MimeTypes.isHeic(mimeType)) {
             // XMP in HEIC motion photos (as taken with a Samsung Camera v12.0.01.50) indicates an `Item:Length` of 68 bytes for the video.
             // This item does not contain the video itself, but only some kind of metadata (no doc, no spec),
@@ -325,15 +342,25 @@ object MultiPage {
         return offsetFromEnd
     }
 
-    fun getTrailerVideoInfo(context: Context, uri: Uri, fileSizeBytes: Long, videoSizeBytes: Long): MediaFormat? {
+    private fun getMotionPhotoVideoInfo(context: Context, uri: Uri, mimeType: String, sizeBytes: Long): MediaFormat? {
+        getMotionPhotoVideoSizing(context, uri, mimeType, sizeBytes)?.let { (videoOffset, videoSize) ->
+            return getEmbedVideoInfo(context, uri, videoOffset, videoSize)
+        }
+        return null
+    }
+
+    fun getTrailerVideoInfo(context: Context, uri: Uri, fileSize: Long, videoSize: Long): MediaFormat? {
+        return getEmbedVideoInfo(context, uri, videoOffset = fileSize - videoSize, videoSize = videoSize)
+    }
+
+    private fun getEmbedVideoInfo(context: Context, uri: Uri, videoOffset: Long, videoSize: Long): MediaFormat? {
         var format: MediaFormat? = null
         val extractor = MediaExtractor()
         var pfd: ParcelFileDescriptor? = null
         try {
-            val videoStartOffset = fileSizeBytes - videoSizeBytes
             pfd = context.contentResolver.openFileDescriptor(uri, "r")
             pfd?.fileDescriptor?.let { fd ->
-                extractor.setDataSource(fd, videoStartOffset, videoSizeBytes)
+                extractor.setDataSource(fd, videoOffset, videoSize)
                 if (extractor.trackCount > 0) {
                     // only consider the first track to represent the appended video
                     val trackIndex = 0
@@ -351,6 +378,36 @@ object MultiPage {
             pfd?.close()
         }
         return format
+    }
+
+    fun getMotionPhotoVideoSizing(context: Context, uri: Uri, mimeType: String, sizeBytes: Long): Pair<Long, Long>? {
+        // default to trailer videos
+        getTrailerVideoSize(context, uri, mimeType, sizeBytes)?.let { videoSize ->
+            val videoOffset = sizeBytes - videoSize
+            return Pair(videoOffset, videoSize)
+        }
+
+        if (MimeTypes.isHeic(mimeType)) {
+            // fallback to video within Samsung SEFD box
+            Mp4ParserHelper.getSamsungSefd(context, uri)?.let { (sefdOffset, bytes) ->
+                val reader = SequentialByteArrayReader(bytes).apply {
+                    isMotorolaByteOrder = false
+                }
+                val start = reader.uInt16
+                val tag = reader.uInt16
+                if (start == 0 && tag == Mp4ParserHelper.SEFD_EMBEDDED_VIDEO_TAG) {
+                    val nameSize = reader.uInt32
+                    val name = reader.getString(nameSize.toInt())
+                    if (name == Mp4ParserHelper.SEFD_MOTION_PHOTO_NAME) {
+                        val videoOffset = sefdOffset + reader.position
+                        val videoSize = reader.available().toLong()
+                        return Pair(videoOffset, videoSize)
+                    }
+                }
+            }
+        }
+
+        return null
     }
 
     fun getTiffPages(context: Context, uri: Uri): ArrayList<FieldMap> {